Moved shapes to separate module, reduced shape boilerplate code.
